% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_GeneClasses.R
\docType{class}
\name{CodingGene-class}
\alias{CodingGene-class}
\alias{CodingGene}
\title{Class CodingGene}
\description{
A class to represent protein-coding genes.
}
\section{Slots}{

\describe{
\item{\code{proteinID}}{A character string representing the protein ID.}

\item{\code{proteinSequence}}{A character string representing the protein sequence.}
}}

\seealso{
\code{\linkS4class{Gene}}
}
