% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Accessors.R
\name{description}
\alias{description}
\alias{description,Gene-method}
\title{Get the description of a Gene}
\usage{
description(x)

\S4method{description}{Gene}(x)
}
\arguments{
\item{x}{A Gene object.}
}
\value{
The description of the gene.
}
\description{
Get the description of a Gene
}
