% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Accessors.R
\name{exon_id}
\alias{exon_id}
\alias{exon_id,GeneStructure-method}
\title{Get the exon ID of a GeneStructure}
\usage{
exon_id(x)

\S4method{exon_id}{GeneStructure}(x)
}
\arguments{
\item{x}{A GeneStructure object.}
}
\value{
The exon ID of the gene structure.
}
\description{
Get the exon ID of a GeneStructure
}
