% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_GeneClasses.R
\docType{class}
\name{Gene-class}
\alias{Gene-class}
\alias{Gene}
\title{Virtual Class Gene}
\description{
A virtual class to represent a general gene structure.
}
\section{Slots}{

\describe{
\item{\code{id}}{A character string representing the gene ID.}

\item{\code{symbol}}{A character string representing the gene symbol.}

\item{\code{name}}{A character string representing the gene name.}

\item{\code{description}}{A character string describing the gene.}

\item{\code{struct}}{A GeneStructure object representing the gene structure.}
}}

\seealso{
\code{\linkS4class{CodingGene}}, \code{\linkS4class{LncRNAGene}}, \code{\linkS4class{MicroRNAGene}}
}
