% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_GeneClasses.R
\name{GeneStructure}
\alias{GeneStructure}
\title{Create a GeneStructure Object}
\usage{
GeneStructure(seqnames, start, end, strand, exon_id)
}
\arguments{
\item{seqnames}{A character vector of sequence names.}

\item{start}{A numeric vector of start positions.}

\item{end}{A numeric vector of end positions.}

\item{strand}{A character vector of strand information.}

\item{exon_id}{A vector of exon IDs.}
}
\value{
A GeneStructure object.
}
\description{
Create a GeneStructure Object
}
\examples{
gene_structure <- GeneStructure(seqnames = "chr1", start = c(100, 230),
                                end = c(200,240), strand = "+",
                                exon_id = 1:2)
gene_structure
}
