% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Accessors.R
\name{id}
\alias{id}
\alias{id,Gene-method}
\title{Get the ID of a Gene}
\usage{
id(x)

\S4method{id}{Gene}(x)
}
\arguments{
\item{x}{A Gene object.}
}
\value{
The ID of the gene.

A character string representing the gene ID.
}
\description{
Get the ID of a Gene
}
\examples{
#I create a GeneStructure Object
gen_struct <- GeneStructure(seqnames = "chr1", start = 100, end = 500,
                            strand = "+", exon_id = 1)

#I consider the class LncRNAGene, which inherited this method from its superclass Gene
gene <- new("LncRNAGene", id = "gene1", symbol = "GENE1", name = "Gene 1",
            description = "A test gene", struct = gen_struct)
id(gene)
}
