% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Accessors.R
\name{lengthProduct}
\alias{lengthProduct}
\alias{lengthProduct,Gene-method}
\title{Get the length of the product sequence of a Gene}
\usage{
lengthProduct(x)

\S4method{lengthProduct}{Gene}(x)
}
\arguments{
\item{x}{A Gene object.}
}
\value{
The length of the product sequence.

An integer representing the length of the product sequence.
}
\description{
Get the length of the product sequence of a Gene
}
\examples{
#I create a GeneStructure Object
gen_struct <- GeneStructure(seqnames = "chr1", start = 100, end = 500,
                            strand = "+", exon_id = 1)
coding_gene <- new("CodingGene", id = "gene1", symbol = "GENE1",
                     name = "Gene 1", description = "A protein-coding gene",
                     struct = gen_struct, proteinID = "P12345",
                     proteinSequence = "MTEYKLVVVG")
lengthProduct(coding_gene)
}
