% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_GeneClasses.R
\docType{class}
\name{LncRNAGene-class}
\alias{LncRNAGene-class}
\alias{LncRNAGene}
\title{Class LncRNAGene}
\description{
A class to represent long non-coding RNA genes.
}
\section{Slots}{

\describe{
\item{\code{lncRNAID}}{A character string representing the lncRNA ID.}

\item{\code{RNASequence}}{A character string representing the RNA sequence.}
}}

\seealso{
\code{\linkS4class{Gene}}
}
