% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_GeneClasses.R
\docType{class}
\name{MicroRNAGene-class}
\alias{MicroRNAGene-class}
\alias{MicroRNAGene}
\title{Class MicroRNAGene}
\description{
A class to represent microRNA genes.
}
\section{Slots}{

\describe{
\item{\code{microRNAID}}{A character string representing the microRNA ID.}

\item{\code{microRNASeedSequence}}{A character string representing the microRNA seed sequence.}
}}

\seealso{
\code{\linkS4class{Gene}}
}
