% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_constructors.R
\name{MicroRNAGene}
\alias{MicroRNAGene}
\title{Constructor function for MicroRNAGene class}
\usage{
MicroRNAGene(ID, HUGO_symbol, gene_name, description, gene_structure, microRNA_ID, microRNA_seed_sequence)
}
\arguments{
\item{ID}{The gene ID (Ensembl ID or NCBI gene ID).}

\item{HUGO_symbol}{The HUGO gene symbol.}

\item{gene_name}{The full gene name.}

\item{description}{A brief description of the gene.}

\item{gene_structure}{The gene structure represented as a GRanges object.}

\item{microRNA_ID}{The microRNA ID.}

\item{microRNA_seed_sequence}{The microRNA seed sequence.}
}
\value{
An instance of the MicroRNAGene class.
}
\description{
This function creates a new instance of the MicroRNAGene class.
}
\examples{
gene_structure <- GenomicRanges::GRanges(seqnames = "chr1", ranges = IRanges::IRanges(5000, 6000), strand = "+")
MicroRNAGene(ID = "ENSG000003", HUGO_symbol = "MIR00001",
             gene_name = "MicroRNA 1",
             description = "MicroRNA",
             gene_structure = gene_structure,
             microRNA_ID = "ENST000005", microRNA_seed_sequence = "UGAGGUAGUAGGUUGUAUGGUAG")
}
\references{
\url{https://www.ensembl.org/index.html?redirect=no}\cr
}
\author{
Niccolò Bianchi\cr UniMi + PoliMI\cr Maintainer: Niccolò
Bianchi\cr E-Mail: \href{mailto:niccolo.bianchi2@studenti.unimi.it}{niccolo.bianchi2@studenti.unimi.it}
}
