% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Accessors.R
\name{name}
\alias{name}
\alias{name,Gene-method}
\title{Get the name of a Gene}
\usage{
name(x)

\S4method{name}{Gene}(x)
}
\arguments{
\item{x}{A Gene object.}
}
\value{
The name of the gene.

A character string representing the gene name.
}
\description{
Get the name of a Gene
}
\examples{
#I create a GeneStructure Object
gen_struct <- GeneStructure(seqnames = "chr1", start = 100, end = 500, strand = "+", exon_id = 1)

#I consider the class MicroRNAGene, which inherited this method from its superclass Gene
gene <- new("MicroRNAGene", id = "gene1", symbol = "GENE1", name = "Gene 1",
           description = "A test gene", struct = gen_struct)
name(gene)
}
