% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Accessors.R
\name{numExons}
\alias{numExons}
\alias{numExons,Gene-method}
\title{Get the number of exons in a Gene}
\usage{
numExons(x)

\S4method{numExons}{Gene}(x)
}
\arguments{
\item{x}{A Gene object.}
}
\value{
The number of exons.
}
\description{
Get the number of exons in a Gene
}
