% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_GeneClasses.R
\docType{class}
\name{PiwiInteractingRNA-class}
\alias{PiwiInteractingRNA-class}
\alias{PiwiInteractingRNA}
\title{Class PiwiInteractingRNA}
\description{
A class to represent piwi-interacting RNA genes.
}
\section{Slots}{

\describe{
\item{\code{piRNAID}}{A character string representing the piRNA ID.}

\item{\code{piRNASequence}}{A character string representing the piRNA sequence.}
}}

\seealso{
\code{\linkS4class{Gene}}
}
