% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Accessors.R
\name{productSequence}
\alias{productSequence}
\alias{productSequence,CodingGene-method}
\alias{productSequence,LncRNAGene-method}
\alias{productSequence,MicroRNAGene-method}
\alias{productSequence,RibosomalRNA-method}
\alias{productSequence,SmallNuclearRNA-method}
\alias{productSequence,PiwiInteractingRNA-method}
\title{Get the product sequence of a Gene}
\usage{
productSequence(x)

\S4method{productSequence}{CodingGene}(x)

\S4method{productSequence}{LncRNAGene}(x)

\S4method{productSequence}{MicroRNAGene}(x)

\S4method{productSequence}{RibosomalRNA}(x)

\S4method{productSequence}{SmallNuclearRNA}(x)

\S4method{productSequence}{PiwiInteractingRNA}(x)
}
\arguments{
\item{x}{A Gene object.}
}
\value{
The product sequence.

A character string representing the product sequence.
}
\description{
Get the product sequence of a Gene
}
\examples{
#I create a GeneStructure Object
gen_struct <- GeneStructure(seqnames = "chr1", start = 100, end = 500,
                            strand = "+", exon_id = 1)
coding_gene <- new("CodingGene", id = "gene1", symbol = "GENE1",
                   name = "Gene 1", description = "A protein-coding gene",
                   struct = gen_struct, proteinID = "P12345",
                   proteinSequence = "MTEYKLVVVG")
productSequence(coding_gene)
}
