% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_GeneClasses.R
\docType{class}
\name{RibosomalRNA-class}
\alias{RibosomalRNA-class}
\alias{RibosomalRNA}
\title{Class RibosomalRNA}
\description{
A class to represent ribosomal RNA genes.
}
\section{Slots}{

\describe{
\item{\code{rRNAID}}{A character string representing the rRNA ID.}

\item{\code{rRNASequence}}{A character string representing the rRNA sequence.}
}}

\seealso{
\code{\linkS4class{Gene}}
}
