% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_GeneClasses.R
\docType{class}
\name{SmallNuclearRNA-class}
\alias{SmallNuclearRNA-class}
\alias{SmallNuclearRNA}
\title{Class SmallNuclearRNA}
\value{
A SmallNuclearRNA instance
}
\description{
A class to represent small nuclear RNA genes.
}
\section{Slots}{

\describe{
\item{\code{snRNAID}}{A character string representing the snRNA ID.}

\item{\code{snRNASequence}}{A character string representing the snRNA sequence.}
}}

\examples{
gen_struct <- GeneStructure(seqnames = "chr1", start = c(100,250),
              end = c(200,300), strand = "+", exon_id = 1:2)
snrna_gene <- new("SmallNuclearRNA", id = "gene5", symbol = "SNR1",
              name = "Small Nuclear RNA 1",
              description = "A small nuclear RNA gene",
              struct = gen_struct, snRNAID = "SNR123",
              snRNASequence = "GCUAGCU")
snrna_gene
}
\seealso{
\code{\linkS4class{Gene}}
}
