% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Accessors.R
\name{struct}
\alias{struct}
\alias{struct,Gene-method}
\title{Get the structure of a Gene}
\usage{
struct(x)

\S4method{struct}{Gene}(x)
}
\arguments{
\item{x}{A Gene object.}
}
\value{
The structure of the gene.
}
\description{
Get the structure of a Gene
}
