% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_GeneClasses.R
\docType{class}
\name{CodingGene-class}
\alias{CodingGene-class}
\alias{CodingGene}
\title{Class CodingGene}
\value{
A CodingGene instance
}
\description{
A class to represent protein-coding genes.
}
\section{Slots}{

\describe{
\item{\code{proteinID}}{A character string representing the protein ID.}

\item{\code{proteinSequence}}{A AAString representing the protein sequence.}
}}

\examples{
library(GenomicRanges)
library(Biostrings)
gen_struct <- GeneStructure(seqnames = "chr1", start = c(100, 250),
                            end = c(200, 300), strand = "+", exon_id = 1:2)
coding_gene <- CodingGene(id = "gene1", symbol = "GENE1",
                   name = "Gene 1",
                   description = "A protein-coding gene",
                   struct = gen_struct, proteinID = "P12345",
                   proteinSequence = "MTEYKLVVVG")
coding_gene
}
\seealso{
\code{\linkS4class{Gene}}
}
