% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Accessors.R
\name{description}
\alias{description}
\alias{description,Gene-method}
\title{Get the description of a Gene}
\usage{
description(x)

\S4method{description}{Gene}(x)
}
\arguments{
\item{x}{A Gene object.}
}
\value{
The description of the gene.

A character string describing the gene.
}
\description{
Get the description of a Gene
}
\examples{
library(GenomicRanges)
gen_struct <- GeneStructure(seqnames = "chr1", start = 100, end = 500,
                            strand = "+", exon_id = 1)
gene <- CodingGene(id = "gene1", symbol = "GENE1", name = "Gene 1",
                   description = "A test gene", struct = gen_struct,
                   proteinID = "P12345", proteinSequence = "MTEYKLVVVG")
description(gene)
}
