% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Accessors.R
\name{id}
\alias{id}
\alias{id,Gene-method}
\title{Get the ID of a Gene}
\usage{
id(x)

\S4method{id}{Gene}(x)
}
\arguments{
\item{x}{A Gene object.}
}
\value{
The ID of the gene.
}
\description{
Get the ID of a Gene
}
