% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Accessors.R
\name{id}
\alias{id}
\alias{id,Gene-method}
\title{Get the ID of a Gene}
\usage{
id(x)

\S4method{id}{Gene}(x)
}
\arguments{
\item{x}{A Gene object.}
}
\value{
The ID of the gene.

A character string representing the gene ID.
}
\description{
Get the ID of a Gene
}
\examples{
library(GenomicRanges)
gen_struct <- GeneStructure(seqnames = "chr1", start = 100, end = 500,
                            strand = "+", exon_id = 1)
gene <- LncRNAGene(id = "gene1", symbol = "GENE1", name = "Gene 1",
                   description = "A test gene", struct = gen_struct,
                   lncRNAID = "LNC1", RNASequence = "AUGCUACG")
id(gene)
}
