% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Accessors.R
\name{lengthProduct}
\alias{lengthProduct}
\alias{lengthProduct,Gene-method}
\title{Get the length of the product sequence of a Gene}
\usage{
lengthProduct(x)

\S4method{lengthProduct}{Gene}(x)
}
\arguments{
\item{x}{A Gene object.}
}
\value{
The length of the product sequence.
}
\description{
Get the length of the product sequence of a Gene
}
