% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_GeneClasses.R
\docType{class}
\name{LncRNAGene-class}
\alias{LncRNAGene-class}
\alias{LncRNAGene}
\title{Class LncRNAGene}
\value{
A LncRNAGene instance
}
\description{
A class to represent long non-coding RNA genes.
}
\section{Slots}{

\describe{
\item{\code{lncRNAID}}{A character string representing the lncRNA ID.}

\item{\code{RNASequence}}{A RNAString representing the RNA sequence.}
}}

\examples{
library(GenomicRanges)
library(Biostrings)
gen_struct <- GeneStructure(seqnames = "chr2", start = c(300, 472),
                            end = c(400, 900), strand = "-", exon_id = 3:4)
lncrna_gene <- LncRNAGene(id = "gene2", symbol = "LNC1",
                   name = "LncRNA 1",
                   description = "A long non-coding RNA gene",
                   struct = gen_struct, lncRNAID = "LNC123",
                   RNASequence = "AUGCUACG")
lncrna_gene
}
\seealso{
\code{\linkS4class{Gene}}
}
