% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_gene.R
\docType{class}
\name{MicroRNAGene-class}
\alias{MicroRNAGene-class}
\title{MicroRNA gene class}
\description{
This is an S4 class to represent a microRNA gene, inheriting from the Gene class.
}
\section{Slots}{

\describe{
\item{\code{microRNA_ID}}{The microRNA ID.}

\item{\code{microRNA_seed_sequence}}{The microRNA seed sequence.}
}}

\examples{
gene_structure <- GenomicRanges::GRanges(seqnames = "chr1", ranges = IRanges::IRanges(start = 5000, end = 6000), strand = "+")
MicroRNAGene(ID = "ENSG000003", HUGO_symbol = "MIR00001",
             gene_name = "MicroRNA 1",
             description = "MicroRNA",
             gene_structure = gene_structure,
             microRNA_ID = "ENST000005", microRNA_seed_sequence = "UGAGGUAGUAGGUUGUAUGGUAG")

}
\references{
\url{https://www.ensembl.org/index.html?redirect=no}\cr
}
\seealso{
\code{\link{ProteinCodingGene}}\cr
\code{\link{LncRNAGene}}\cr
}
\author{
Niccolò Bianchi\cr UniMi + PoliMI\cr Maintainer: Niccolò
Bianchi\cr E-Mail: \href{mailto:niccolo.bianchi2@studenti.unimi.it}{niccolo.bianchi2@studenti.unimi.it}
}
