% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Accessors.R
\name{name}
\alias{name}
\alias{name,Gene-method}
\title{Get the name of a Gene}
\usage{
name(x)

\S4method{name}{Gene}(x)
}
\arguments{
\item{x}{A Gene object.}
}
\value{
The name of the gene.

A character string representing the gene name.
}
\description{
Get the name of a Gene
}
\examples{
library(GenomicRanges)
gen_struct <- GeneStructure(seqnames = "chr1", start = 100, end = 500,
                            strand = "+", exon_id = 1)
gene <- MicroRNAGene(id = "gene1", symbol = "GENE1", name = "Gene 1",
                     description = "A test gene", struct = gen_struct,
                     microRNAID = "MIR1", microRNASeedSequence = "UAGCUU")
name(gene)
}
