% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Accessors.R
\name{name}
\alias{name}
\alias{name,Gene-method}
\title{Get the name of a Gene}
\usage{
name(x)

\S4method{name}{Gene}(x)
}
\arguments{
\item{x}{A Gene object.}
}
\value{
The name of the gene.
}
\description{
Get the name of a Gene
}
