% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_GeneClasses.R
\docType{class}
\name{PiwiInteractingRNA-class}
\alias{PiwiInteractingRNA-class}
\alias{PiwiInteractingRNA}
\title{Class PiwiInteractingRNA}
\value{
A PiwiInteractingRNA instance
}
\description{
A class to represent piwi-interacting RNA genes.
}
\section{Slots}{

\describe{
\item{\code{piRNAID}}{A character string representing the piRNA ID.}

\item{\code{piRNASequence}}{A RNAString representing the piRNA sequence.}
}}

\examples{
library(GenomicRanges)
library(Biostrings)
gen_struct <- GeneStructure(seqnames = "chr1", start = c(100, 250),
                            end = c(200, 300), strand = "+", exon_id = 1:2)
pirna_gene <- PiwiInteractingRNA(id = "gene6", symbol = "PIWI1",
                  name = "Piwi-Interacting RNA 1",
                  description = "A piwi-interacting RNA gene",
                  struct = gen_struct, piRNAID = "PIWI123",
                  piRNASequence = "UCGUAU")
pirna_gene
}
\seealso{
\code{\linkS4class{Gene}}
}
