% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Accessors.R
\name{productSequence}
\alias{productSequence}
\alias{productSequence,CodingGene-method}
\alias{productSequence,LncRNAGene-method}
\alias{productSequence,MicroRNAGene-method}
\alias{productSequence,RibosomalRNA-method}
\alias{productSequence,SmallNuclearRNA-method}
\alias{productSequence,PiwiInteractingRNA-method}
\title{Get the product sequence of a Gene}
\usage{
productSequence(x)

\S4method{productSequence}{CodingGene}(x)

\S4method{productSequence}{LncRNAGene}(x)

\S4method{productSequence}{MicroRNAGene}(x)

\S4method{productSequence}{RibosomalRNA}(x)

\S4method{productSequence}{SmallNuclearRNA}(x)

\S4method{productSequence}{PiwiInteractingRNA}(x)
}
\arguments{
\item{x}{A Gene object.}
}
\value{
The product sequence.
}
\description{
Get the product sequence of a Gene
}
