% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_GeneClasses.R
\docType{class}
\name{SmallNuclearRNA-class}
\alias{SmallNuclearRNA-class}
\alias{SmallNuclearRNA}
\title{Class SmallNuclearRNA}
\description{
A class to represent small nuclear RNA genes.
}
\section{Slots}{

\describe{
\item{\code{snRNAID}}{A character string representing the snRNA ID.}

\item{\code{snRNASequence}}{A character string representing the snRNA sequence.}
}}

\seealso{
\code{\linkS4class{Gene}}
}
