% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Accessors.R
\name{struct}
\alias{struct}
\alias{struct,Gene-method}
\title{Get the structure of a Gene}
\usage{
struct(x)

\S4method{struct}{Gene}(x)
}
\arguments{
\item{x}{A Gene object.}
}
\value{
The structure of the gene.

A GeneStructure object representing the gene structure.
}
\description{
Get the structure of a Gene
}
\examples{
library(GenomicRanges)
gen_struct <- GeneStructure(seqnames = "chr1", start = 100, end = 500,
                            strand = "+", exon_id = 1)
gene <- RibosomalRNA(id = "gene1", symbol = "GENE1",
                     name = "Gene 1", description = "A test gene",
                     struct = gen_struct, rRNAID = "RIB1",
                     rRNASequence = "AGGCUAG")
struct(gene)
}
