% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Accessors.R
\name{symbol}
\alias{symbol}
\alias{symbol,Gene-method}
\title{Get the symbol of a Gene}
\usage{
symbol(x)

\S4method{symbol}{Gene}(x)
}
\arguments{
\item{x}{A Gene object.}
}
\value{
The symbol of the gene.
}
\description{
Get the symbol of a Gene
}
