% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Accessors.R
\name{symbol}
\alias{symbol}
\alias{symbol,Gene-method}
\title{Get the symbol of a Gene}
\usage{
symbol(x)

\S4method{symbol}{Gene}(x)
}
\arguments{
\item{x}{A Gene object.}
}
\value{
The symbol of the gene.

A character string representing the gene symbol.
}
\description{
Get the symbol of a Gene
}
\examples{
library(GenomicRanges)
gen_struct <- GeneStructure(seqnames = "chr1", start = 100, end = 500,
                            strand = "+", exon_id = 1)
gene <- CodingGene(id = "gene1", symbol = "GENE1", name = "Gene 1",
                   description = "A test gene", struct = gen_struct,
                   proteinID = "P12345", proteinSequence = "MTEYKLVVVG")
symbol(gene)
}
